Split main.py into parser.py
